Added Pull Request Template (#3)
## Description
<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->
<!-- If there is a ticket for this work item, including Github issues,
please include it here as well i.e. `[<Task Name>](www.story-link.com)`
-->
I have added a template for Pull Requests. So when a developer creates a
PR the subsequent commit message to 'main' will have a consistent
message containing the details around what was added/modified with that
particular commit.

### Type of Change
<!-- Please delete options that are not relevant. -->
- [x] Other (please describe): This is a Github update, it doesn't
effect code.

## Does this introduce a breaking change?
<!-- If this introduces a breaking change, please add the **Breaking**
tag to the PR. -->

- [ ] Yes
- [x] No

## Checklist:
<!-- Please ensure the following items are taken care of as part of your
PR. If they are not, your PR will NOT be approved. -->

- [x] Doc strings have been added to any public APIs
- [x] Appropriate tests have been added
- [x] The CHANGELOG.md file has been updated with your changes (1 entry
per PR)
- [x] Before merging, rebase your branch with the target branch