Dashboard Screen (#4)
## Description
<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->
<!-- If there is a ticket for this work item, including Github issues,
please include it here as well i.e. `[<Task Name>](www.story-link.com)`
-->

The creation of the Dashboard screen. This will be the screen the users
will first see when they open the app, whether they are logged in or
not.

This screen will control the majority of navigation for the app. It is a
container which will contain the main screens within the app as it is
developed.

### Type of Change
<!-- Please delete options that are not relevant. -->
- [x] Feature

## Does this introduce a breaking change?
<!-- If this introduces a breaking change, please add the **Breaking**
tag to the PR. -->

- [ ] Yes
- [x] No

## How Can This PR Be Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce. Please also list any relevant
details for your test configuration. -->
<!-- Please include screenshots/video where applicable. -->

**Test Configuration**:
* Firmware version: 17.5.1
* Hardware: iPhone 15 Pro Max
* Is Simulator:
  - [ ] Yes
  - [x] No

In the application test the following scenarios:
1) Swipe the main portion (the white space with the text object in the
middle) of the screen from left to right.
* This should update the ScrollingTabView at the top of the screen in
accordance with the currently selected view.
2) Scroll to multiple different positions in the ScrollingTabView,
tapping on the options throughout.
* This should update the main view with the relevant screen. At present
its just a Text view in the middle with the associated tab title.
3) Select a tab option that is not the "Dashboard" option, then tap on
the ChurchButton (in the top-leading corner)
* This should navigate you back to the "Dashboard" ScrollingTabView
option along with updating the main view accordingly

## Checklist:
<!-- Please ensure the following items are taken care of as part of your
PR. If they are not, your PR will NOT be approved. -->

- [x] Doc strings have been added to any public APIs
- [x] Appropriate tests have been added
- [x] The CHANGELOG.md file has been updated with your changes (1 entry
per PR)
- [x] Before merging, rebase your branch with the target branch